GenderDetail trait

+++ src/personal_info/traits/base_state.rs
@@ -1,18 +1,18 @@
 use crate::{PKError, PKResult};
 
 pub trait BaseState {
-    fn hp(&self) -> i32;
-    fn atk(&self) -> i32;
-    fn def(&self) -> i32;
-    fn spa(&self) -> i32;
-    fn spd(&self) -> i32;
-    fn spe(&self) -> i32;
+    fn hp(&self) -> u16;
+    fn atk(&self) -> u16;
+    fn def(&self) -> u16;
+    fn spa(&self) -> u16;
+    fn spd(&self) -> u16;
+    fn spe(&self) -> u16;
 
-    fn get_base_state_total(&self) -> i32 {
+    fn get_base_state_total(&self) -> u16 {
         self.hp() + self.atk() + self.def() + self.spa() + self.spd() + self.spe()
     }
 
-    fn get_base_state_value(&self, index: usize) -> PKResult<i32> {
+    fn get_base_state_value(&self, index: usize) -> PKResult<u16> {
         match index {
             0 => Ok(self.hp()),
             1 => Ok(self.atk()),
@@ -24,7 +24,7 @@ pub trait BaseState {
         }
     }
 
-    fn get_sorted_stat_indices(&self, stats: &mut [(usize, i32)]) -> PKResult<()> {
+    fn get_sorted_stat_indices(&self, stats: &mut [(usize, u16)]) -> PKResult<()> {
         for (i, stat) in stats.iter_mut().enumerate() {
             *stat = (i, self.get_base_state_value(i)?);
         }
