+++ src/personal_info/traits/effort_value_yield.rs
@@ -1,8 +1,8 @@
 pub trait EffortValueYield {
-    fn ev_hp(&self) -> i32;
-    fn ev_atk(&self) -> i32;
-    fn ev_def(&self) -> i32;
-    fn ev_spe(&self) -> i32;
-    fn ev_spa(&self) -> i32;
-    fn ev_spd(&self) -> i32;
+    fn ev_hp(&self) -> u8;
+    fn ev_atk(&self) -> u8;
+    fn ev_def(&self) -> u8;
+    fn ev_spe(&self) -> u8;
+    fn ev_spa(&self) -> u8;
+    fn ev_spd(&self) -> u8;
 }
